refactor: 파일 구조 변경
파일 구조 변경된 부분은 merge했다.
