--- backend/src/main/java/dev/chrono/chronochallenge/member/service/QuizService.java
@@ -1,6 +1,0 @@
-package dev.chrono.chronochallenge.service;
-
-import dev.chrono.chronochallenge.model.Quiz;
-
-public interface QuizService extends CrudSevice<Quiz, Long>{
-}
