+++ backend/src/main/java/dev/chrono/chronochallenge/quiz/service/QuizService.java
@@ -1,0 +1,7 @@
+package dev.chrono.chronochallenge.quiz.service;
+
+import dev.chrono.chronochallenge.common.service.CrudSevice;
+import dev.chrono.chronochallenge.quiz.model.Quiz;
+
+public interface QuizService extends CrudSevice<Quiz, Long> {
+}
